requirements.txt
absl-py==0.8.0
alabaster==0.7.12
argh==0.26.2
asgiref==3.2.3
astor==0.8.0
audioread==2.1.8
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.9.0
bitsets==0.7.15
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
chromaprint==0.5
Click==7.0
colorama==0.4.1
comtypes==1.1.7
concepts==0.9
configparser==4.0.2
confusable-homoglyphs==3.2.0
cycler==0.10.0
darr==0.2.2
dcase-util==0.2.10
decorator==4.4.0
Django==3.0.1
django-crispy-forms==1.8.1
django-crispy-forms-registration==0.1.3
django-registration==3.0.1
docker-pycreds==0.4.0
docutils==0.15.2
enum34==1.1.10
features==0.5.11
ffmpeg==1.4
ffmpeg-python==0.2.0
fileconfig==0.5.9
future==0.17.1
fuzzywuzzy==0.17.0
gast==0.2.2
gitdb2==2.0.6
GitPython==3.0.5
google-auth==1.8.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
gql==0.1.0
graphql-core==2.2.1
graphviz==0.13
grpcio==1.26.0
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
helpdev==0.6.10
helpers==0.2.0
idna==2.8
image==1.5.27
imagesize==1.1.0
importlib-metadata==0.23
imutils==0.5.3
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.7.0
llvmlite==0.29.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.2.0
MouseInfo==0.1.3
numba==0.45.1
numpy==1.17.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.1
parso==0.5.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow==6.2.1
plotly==4.4.1
pocketsphinx==0.1.15
presets==0.1.3
promise==2.2.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.3
pyacoustid==1.1.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAudio==0.2.11
PyAutoGUI==0.9.50
pycaw==20181226
pycparser==2.19
pydot-ng==2.0.0
pydub==0.23.1
PyGetWindow==0.0.8
Pygments==2.4.2
PyMsgBox==1.0.8
pyparsing==2.4.2
pyperclip==1.8.0
pypiwin32==223
PyQt5==5.13.2
PyQt5-sip==12.7.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.0
python-magic==0.4.15
python-speech-features==0.6
pyttsx==1.1
pyttsx3==2.71
PyTweening==1.0.3
pytz==2019.2
pywin32==227
PyYAML==5.1.2
QDarkStyle==2.7
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.0
Rx==1.6.1
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
sentry-sdk==0.13.2
shortuuid==0.5.0
simpleaudio==1.0.2
six==1.12.0
smmap2==2.0.5
snowballstemmer==2.0.0
sound==0.1.0
SoundFile==0.10.2
soupsieve==2.0
SpeechRecognition==3.8.1
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
split-folders==0.3.1
sqlparse==0.3.0
subprocess32==3.5.4
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tqdm==4.41.0
traitlets==4.3.3
urllib3==1.25.6
validators==0.14.1
voxpopuli==0.3.3
wandb==0.8.15
watchdog==0.9.0
wcwidth==0.1.7
Werkzeug==0.16.0
wit==5.1.0
wrapt==1.11.2
zipp==0.6.0
